+++ tool/backup.js
@@ -1,22 +1,17 @@
 #!/usr/bin/env node
 'use strict';
+var fs = require('fs');
+var Database = require('../bot/db.js');
+var settings = require('../private/settings.js');
+var svcAcct = '../private/serviceAccountCredentials.json';
+var BASEURL = settings.FIREBASE.BASEURL;
+var db = new Database(svcAcct, BASEURL);
 
 /*************************************
   Back up our DB via Cron.  Mainly
   to backup triggers
 */
 
-var firebase = require('firebase');
-var settings = require('../private/settings.js');
-var fbCreds  = require('../private/serviceAccountCredentials.json');
-var fs = require('fs');
-
-firebase.initializeApp({
-  serviceAccount: fbCreds,
-  databaseURL: settings.FIREBASE.BASEURL
-});
-
-var db = firebase.database();
 
 db.ref().once('value', function(snapshot) {
   var val = snapshot.val();